ACCEPT WEB PORTAL
The ACCEPT Web Portal is a .Net MVC4 application, all data is consumed
from the ACCEPT API endpoint, so as part of the configuration process, a
valid ACCEPT API endpoint needs to be setup in the Web.config file. Note
there are more configuration settings that need to be
provided(application settings section within Web.config).